% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveArchiveLinks.R
\name{retrieveArchiveLinks}
\alias{retrieveArchiveLinks}
\title{retrieveArchiveLinks: Retrieving Links of Lower-level web pages of mementos from the Internet Archive}
\usage{
retrieveArchiveLinks(ArchiveUrls, encoding = "UTF-8")
}
\arguments{
\item{ArchiveUrls}{A string of the memento of the Internet Archive}

\item{encoding}{Specify a default encoding for the homepage. Default is 'UTF-8'}
}
\value{
This function retrieves the links of all lower-level web pages of mementos of a homepage available from the Internet Archive. It returns a tibble including the baseUrl and all links of lower-level web pages. However, a memento being stored in the Internet Archive does not guarantee that the information from the homepage can be actually scraped.
}
\description{
\code{retrieveArchiveLinks} retrieves the Urls of mementos stored in the Internet Archive
}
\examples{
\dontrun{
retrieveArchiveLinks("http://web.archive.org/web/20190801001228/https://www.spiegel.de/")
}
}
