% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeArchiveUrls.R
\name{scrapeArchiveUrls}
\alias{scrapeArchiveUrls}
\title{scrapeArchiveUrls: Scraping Urls from the Internet Archive}
\usage{
scrapeArchiveUrls(
  Urls,
  Paths,
  startnum = 1,
  attachto = NaN,
  CSS = F,
  archiveDate = F,
  ignoreErrors = F,
  stopatempty = T,
  emptylim = 10,
  encoding = "UTF-8"
)
}
\arguments{
\item{Urls}{A string of the memento of the Internet Archive}

\item{Paths}{XXX}

\item{startnum}{XXX}

\item{attachto}{XXX}

\item{CSS}{XXX}

\item{archiveDate}{XXX}

\item{ignoreErrors}{XXX}

\item{stopatempty}{XXX}

\item{emptylim}{XXX}

\item{encoding}{Specify a default encoding for the homepage. Default is 'UTF-8'}
}
\value{
This function scrapes the content of mementos or lower-level web pages from the Internet Archive. It returns a tibble including Urls and the scraped content. However, a memento being stored in the Internet Archive does not guarantee that the information from the homepage can be actually scraped.
}
\description{
\code{scrapeArchiveUrls} scrapes Urls of mementos and lower-level web pages stored in the Internet Archive using XPaths as default
}
