% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveOverview.R
\name{archiveOverview}
\alias{archiveOverview}
\title{archiveOverview: Getting a first glimpse of mementos available in the Internet Archive}
\usage{
archiveOverview(homepage, startDate, endDate)
}
\description{
\code{archiveOverview} provides an overview of available mementos of the homepage from the Internet Archive
}
\details{
This functions provides an overview of mementos available from the Internet Archive. It returns a calender indicating all dates in which mementos of the homepage have been stored in the Internet Archive at least once. However, a memento being stored in the Internet Archive does not guarantee that the information from the homepage can be actually scraped.
}
