% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveOverview.R
\name{archiveOverview}
\alias{archiveOverview}
\title{archiveOverview: Getting a first glimpse of mementos available in the Internet Archive}
\usage{
archiveOverview(homepage, startDate, endDate)
}
\arguments{
\item{homepage}{A character vector of the homepage, including the top-level-domain}

\item{startDate}{A character vector of the starting date of the overview. Accepts a large variety of date formats (see \link[anytime]{anytime})}

\item{endDate}{A character vector of the ending date of the overview. Accepts a large variety of date formats (see \link[anytime]{anytime})}
}
\value{
This function provides an overview of mementos available from the Internet Archive. It returns a calender indicating all dates in which mementos of the homepage have been stored in the Internet Archive at least once. However, a memento being stored in the Internet Archive does not guarantee that the information from the homepage can be actually scraped.
}
\description{
\code{archiveOverview} provides an overview of available mementos of the homepage from the Internet Archive
}
\examples{
\dontrun{
archiveOverview(homepage = "www.spiegel.de", startDate = "20180601", endDate = "20190615")
archiveOverview(homepage = "nytimes.com", startDate = "2018-06-01", endDate = "2019-05-01")
}
}
